altered outputs folder anticipating Guillame's code. Tweaked about.html
