+++ templates/about.html
@@ -4,11 +4,13 @@
 
 <h2>{{ title }}</h2>
 <h3>{{ message }}</h3>
-
+<p>
+    Required python modules to install: <ul>flask</ul> <ul>flask-websocket</ul> <ul>eventlet (or gevent + gevent-websocket)</ul><br>All other code we wrote can be found on <a href="https://github.com/DVC-Viking-Robotics/webapp">GitHub</a>
+</p>
 <p id="addresses">
-    This website's Support: <a href="mailto:2bndy5@gmail.com">Brendan Doherty</a>
+    This website's Support: <a href="https://github.com/DVC-Viking-Robotics/webapp">DVC Viking Robotics</a>
     <br />
-    Marketing: <a href="mailto:dvcrobotics@gmail.com">Diablo Valley Community College - Robotics Club</a>
+    Marketing: Contact us! <a href="mailto:dvcrobotics@gmail.com">Diablo Valley Community College - Robotics Club</a>
 </p>
 
 <p id="placeholer">
